Bug 1856913 - Refactor ReviewQualityCheckVendorsService and repackage ext
