+++ fenix/app/src/main/java/org/mozilla/fenix/shopping/ui/ext/ProductVendor.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.shopping.ext
+package org.mozilla.fenix.shopping.ui.ext
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
