+++ fenix/app/src/test/java/org/mozilla/fenix/shopping/fake/FakeReviewQualityCheckVendorsService.kt
@@ -15,6 +15,4 @@ class FakeReviewQualityCheckVendorsService(
     ),
 ) : ReviewQualityCheckVendorsService {
     override fun productVendors(): List<ProductVendor> = productVendors
-
-    override fun productVendor(): ProductVendor = productVendors().first()
 }
